backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Caching
redis==5.0.1

# OpenAI Integration
openai==1.3.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-json-logger==2.0.7